+++ tests/rx/function_group.cpp
@@ -1,7 +1,7 @@
 #include "rx_test.hpp"
 
 TEST_F(RxTest, function_group_f4_f0) {
-  auto state{RandomInterval(0b0'0000u, 0b1'1111u)};
+  auto state{RandomInterval<uint8_t>(0b0'0000u, 0b1'1111u)};
   Expectation function{EXPECT_CALL(_mock, function(3u, 0b11111u, state))};
   Receive(dcc::make_function_group_f4_f0_packet(3u, state));
   Execute();
@@ -17,17 +17,17 @@ TEST_F(RxTest, function_group_f4_f0_exception) {
 }
 
 TEST_F(RxTest, function_group_f8_f5) {
-  auto state{RandomInterval(0x0u, 0xFu)};
-  Expectation function{
-    EXPECT_CALL(_mock, function(3u, 0xFu << 5u, state << 5u))};
+  auto state{RandomInterval<uint8_t>(0x0u, 0xFu)};
+  Expectation function{EXPECT_CALL(
+    _mock, function(3u, 0xFu << 5u, static_cast<uint32_t>(state << 5u)))};
   Receive(dcc::make_function_group_f8_f5_packet(3u, state));
   Execute();
 }
 
 TEST_F(RxTest, function_group_f12_f9) {
-  auto state{RandomInterval(0x0u, 0xFu)};
-  Expectation function{
-    EXPECT_CALL(_mock, function(3u, 0xFu << 9u, state << 9u))};
+  auto state{RandomInterval<uint8_t>(0x0u, 0xFu)};
+  Expectation function{EXPECT_CALL(
+    _mock, function(3u, 0xFu << 9u, static_cast<uint32_t>(state << 9u)))};
   Receive(dcc::make_function_group_f12_f9_packet(3u, state));
   Execute();
 }